+++ docs/changelog.md
@@ -1,8 +1,9 @@
 # Changelog
 
-## 1.14.2 (unreleased to Maven Central)
+## 1.15.0 (unreleased to Maven Central)
 
 - Adds documentation section filenames into the model.
+- Adds support for custom elements on dynamic views.
 
 ## 1.14.1 (15th August 2022)